Plot the residuals of the binned ADC samples

+++ analysis/scum/radio/tuning/tuning_codes_main.py
@@ -16,7 +16,7 @@ TX_LABEL = "TX"
 RX_LABEL = "RX"
 
 
-def _tuning_code_formatter(x: int, position: float) -> str:
+def _tuning_code_formatter(x: float, position: float) -> str:
     """Formats the tuning code tick labels.
 
     Args:
